Refactor fs monitoring module: move to separate package, minimizing dependencies
Also slightly refactored network.go for consistency
